Refactor dynamic canvas into MVC modules
